requirements.txt.txt
python-telegram-bot==22.0
gspread==6.1.2
google-auth==2.34.0
